requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
Pillow
seaborn
tkinter
